Updating file structure for better developing
